Move form builder helpers to forms module
