+++ collect_app/src/test/java/org/odk/collect/android/formmanagement/ServerFormsSynchronizerTest.java
@@ -7,7 +7,7 @@ import org.odk.collect.forms.FormSourceException;
 import org.odk.collect.forms.FormsRepository;
 import org.odk.collect.forms.instances.InstancesRepository;
 import org.odk.collect.android.itemsets.FastExternalItemsetsRepository;
-import org.odk.collect.android.support.FormUtils;
+import org.odk.collect.forms.testsupport.FormUtils;
 import org.odk.collect.android.support.InMemFormsRepository;
 import org.odk.collect.android.support.InMemInstancesRepository;
