+++ collect_app/src/test/java/org/odk/collect/android/instancemanagement/InstanceDeleterTest.java
@@ -4,7 +4,7 @@ import org.junit.Test;
 import org.odk.collect.forms.Form;
 import org.odk.collect.forms.FormsRepository;
 import org.odk.collect.forms.instances.Instance;
-import org.odk.collect.android.support.FormUtils;
+import org.odk.collect.forms.testsupport.FormUtils;
 import org.odk.collect.android.support.InMemFormsRepository;
 import org.odk.collect.android.support.InMemInstancesRepository;
 import org.odk.collect.testshared.TempFiles;
