+++ collect_app/src/test/java/org/odk/collect/android/provider/FormsProviderTest.java
@@ -16,7 +16,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.odk.collect.android.fastexternalitemset.ItemsetDbAdapter;
 import org.odk.collect.android.storage.StorageInitializer;
-import org.odk.collect.android.support.FormUtils;
+import org.odk.collect.forms.testsupport.FormUtils;
 import org.odk.collect.android.utilities.FileUtils;
 import org.odk.collect.shared.Md5;
 import org.robolectric.shadows.ShadowEnvironment;
