+++ forms/build.gradle
@@ -3,6 +3,8 @@ plugins {
     id 'kotlin'
 }
 
+apply from: '../config/quality.gradle'
+
 java {
     sourceCompatibility = JavaVersion.VERSION_1_7
     targetCompatibility = JavaVersion.VERSION_1_7
@@ -10,4 +12,5 @@ java {
 
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    implementation "commons-io:commons-io:2.5"
 }
